backend/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.7
Django==3.2.18
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
mysqlclient==2.2.1
openai==1.7.1
packaging==23.2
protobuf==4.21.9
pycparser==2.21
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
